Edit call-to-action-section
added a blue background to the call-to-action-container.
added padding to the call-to-action-container.
added justify-content to provide space between the two smaller divs.
added align-content center, so the button didn't fill the cross axis.
fixed the call-to-action-button from not working by changing it from
a class to an id in style.css.
added a white border to the call-to-action-button.
created separate classes for the call-to-action-text.
edited each call to action text to have different sizes and colors.